+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/actions/generators/proposalActionTokenMint.ts
@@ -17,9 +17,11 @@ export const generateProposalActionTokenMint = (
             { name: 'tokenAmount', type: 'string', value: '2000000000000000000' },
         ],
     },
-    receivers: [],
-    tokenSupply: 10000,
-    holdersCount: 500,
+    receiver: {
+        currentBalance: '0',
+        newBalance: '5',
+        address: '0x32c2FE388ABbB3e678D44DF6a0471086D705316a',
+    },
     tokenSymbol: 'PDC',
     ...action,
 });