+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/actions/proposalActionTokenMint/proposalActionTokenMint.stories.tsx
@@ -23,23 +23,11 @@ export const Default: Story = {
         return (
             <ProposalActionTokenMint
                 action={generateProposalActionTokenMint({
-                    receivers: [
-                        {
-                            currentBalance: 0,
-                            newBalance: 5,
-                            address: '0x32c2FE388ABbB3e678D44DF6a0471086D705316a',
-                        },
-                        {
-                            currentBalance: 100,
-                            newBalance: 110,
-                            address: '0xeefB13C7D42eFCc655E528dA6d6F7bBcf9A2251d',
-                        },
-                        {
-                            currentBalance: 0,
-                            newBalance: 200,
-                            address: '0x742d35Cc6634C0532925a3b844Bc454e4438f44e',
-                        },
-                    ],
+                    receiver: {
+                        currentBalance: '0',
+                        newBalance: '5',
+                        address: '0x32c2FE388ABbB3e678D44DF6a0471086D705316a',
+                    },
                 })}
             />
         );